CLN: move diff, take to algorithms.py
