+++ pandas/sparse/frame.py
@@ -16,6 +16,7 @@ from pandas.core.series import Series
 from pandas.core.frame import (DataFrame, extract_index, _prep_ndarray,
                                _default_index)
 import pandas.core.common as com
+import pandas.core.algorithms as algos
 from pandas.core.internals import (BlockManager,
                                    create_block_manager_from_arrays)
 from pandas.core.generic import NDFrame
@@ -593,9 +594,9 @@ class SparseDataFrame(DataFrame):
             if col not in self:
                 continue
             if row_indexer is not None:
-                new_arrays[col] = com.take_1d(self[col].get_values(),
-                                              row_indexer,
-                                              fill_value=fill_value)
+                new_arrays[col] = algos.take_1d(self[col].get_values(),
+                                                row_indexer,
+                                                fill_value=fill_value)
             else:
                 new_arrays[col] = self[col]
